Big hint/reveal change
Hint 1/2 will now show "Capital: XXX" or "Country: XXX" under the flag depending on the question. If no flag is included in the question (type: Country -> Flag) then the answers will get all capitals or countries in front of their flags depending again what is the question for.

Hint 1/2 will take 0.75 point
Hint 2/2 will take 0.75 point
Reveal will take 0.5 point.

+++ app/src/main/java/com/example/flaggameandroid/persistence/PersistedModels.kt
@@ -14,7 +14,7 @@ import com.example.flaggameandroid.feature.app.AppLanguage
 data class PersistedAppState(
   val hintDifficulty: HintDifficulty = HintDifficulty.Medium,
   val reminderEnabled: Boolean = true,
-  val hintCount: Int = 0,
+  val hintCount: Double = 0.0,
   val level: Int = 1,
   val hintsTowardNextLevel: Int = 0,
   val correctAnswersTowardNextLevel: Int = 0,
